refactor(start-wrt): inline the router-port confirm dialog into its only caller
Review feedback on #3634: confirmRouterPortOverride had a single call site
and PublishedPortsService.save() already funnels every UI path through it,
so it lives in the service now — dialogs/i18n come from the service's own
injections instead of parameters.
